# [core] use factory methods for registering services