--- a/src/Controls/src/Xaml/Hosting/AppHostBuilderExtensions.cs
+++ b/src/Controls/src/Xaml/Hosting/AppHostBuilderExtensions.cs
@@ -62,101 +62,115 @@ namespace Microsoft.Maui.Controls.Hosting
 
 		public static IMauiHandlersCollection AddMauiControlsHandlers(this IMauiHandlersCollection handlersCollection)
 		{
-			handlersCollection.AddHandler<CollectionView, CollectionViewHandler>();
-			handlersCollection.AddHandler<CarouselView, CarouselViewHandler>();
-			handlersCollection.AddHandler<Application, ApplicationHandler>();
-			handlersCollection.AddHandler<ActivityIndicator, ActivityIndicatorHandler>();
-			handlersCollection.AddHandler<BoxView, BoxViewHandler>();
-			handlersCollection.AddHandler<Button, ButtonHandler>();
-			handlersCollection.AddHandler<CheckBox, CheckBoxHandler>();
-			handlersCollection.AddHandler<DatePicker, DatePickerHandler>();
-			handlersCollection.AddHandler<Editor, EditorHandler>();
-			handlersCollection.AddHandler<Entry, EntryHandler>();
-			handlersCollection.AddHandler<GraphicsView, GraphicsViewHandler>();
-			handlersCollection.AddHandler<Image, ImageHandler>();
-			handlersCollection.AddHandler<Label, LabelHandler>();
-			handlersCollection.AddHandler<Layout, LayoutHandler>();
-			handlersCollection.AddHandler<Picker, PickerHandler>();
-			handlersCollection.AddHandler<ProgressBar, ProgressBarHandler>();
-			handlersCollection.AddHandler<ScrollView, ScrollViewHandler>();
-			handlersCollection.AddHandler<SearchBar, SearchBarHandler>();
-			handlersCollection.AddHandler<Slider, SliderHandler>();
-			handlersCollection.AddHandler<Stepper, StepperHandler>();
-			handlersCollection.AddHandler<Switch, SwitchHandler>();
-			handlersCollection.AddHandler<TimePicker, TimePickerHandler>();
-			handlersCollection.AddHandler<Page, PageHandler>();
-			handlersCollection.AddHandler<WebView, WebViewHandler>();
-			handlersCollection.AddHandler<Border, BorderHandler>();
-			handlersCollection.AddHandler<IContentView, ContentViewHandler>();
-			handlersCollection.AddHandler<Shapes.Ellipse, ShapeViewHandler>();
-			handlersCollection.AddHandler<Shapes.Line, LineHandler>();
-			handlersCollection.AddHandler<Shapes.Path, PathHandler>();
-			handlersCollection.AddHandler<Shapes.Polygon, PolygonHandler>();
-			handlersCollection.AddHandler<Shapes.Polyline, PolylineHandler>();
-			handlersCollection.AddHandler<Shapes.Rectangle, RectangleHandler>();
-			handlersCollection.AddHandler<Shapes.RoundRectangle, RoundRectangleHandler>();
-			handlersCollection.AddHandler<Window, WindowHandler>();
-			handlersCollection.AddHandler<ImageButton, ImageButtonHandler>();
-			handlersCollection.AddHandler<IndicatorView, IndicatorViewHandler>();
-			handlersCollection.AddHandler<RadioButton, RadioButtonHandler>();
-			handlersCollection.AddHandler<RefreshView, RefreshViewHandler>();
-			handlersCollection.AddHandler<SwipeItem, SwipeItemMenuItemHandler>();
-			handlersCollection.AddHandler<SwipeView, SwipeViewHandler>();
+			handlersCollection.AddHandler<CollectionView>(_ => new CollectionViewHandler());
+			handlersCollection.AddHandler<CarouselView>(_ => new CarouselViewHandler());
+			handlersCollection.AddHandler<Application>(_ => new ApplicationHandler());
+			handlersCollection.AddHandler<ActivityIndicator>(_ => new ActivityIndicatorHandler());
+			handlersCollection.AddHandler<BoxView>(_ => new BoxViewHandler());
+			handlersCollection.AddHandler<Button>(_ => new ButtonHandler());
+			handlersCollection.AddHandler<CheckBox>(_ => new CheckBoxHandler());
+			handlersCollection.AddHandler<DatePicker>(_ => new DatePickerHandler());
+			handlersCollection.AddHandler<Editor>(_ => new EditorHandler());
+			handlersCollection.AddHandler<Entry>(_ => new EntryHandler());
+			handlersCollection.AddHandler<GraphicsView>(_ => new GraphicsViewHandler());
+			handlersCollection.AddHandler<Image>(_ => new ImageHandler());
+			handlersCollection.AddHandler<Label>(_ => new LabelHandler());
+			handlersCollection.AddHandler<Layout>(_ => new LayoutHandler());
+			handlersCollection.AddHandler<Picker>(_ => new PickerHandler());
+			handlersCollection.AddHandler<ProgressBar>(_ => new ProgressBarHandler());
+			handlersCollection.AddHandler<ScrollView>(_ => new ScrollViewHandler());
+			handlersCollection.AddHandler<SearchBar>(_ => new SearchBarHandler());
+			handlersCollection.AddHandler<Slider>(_ => new SliderHandler());
+			handlersCollection.AddHandler<Stepper>(_ => new StepperHandler());
+			handlersCollection.AddHandler<Switch>(_ => new SwitchHandler());
+			handlersCollection.AddHandler<TimePicker>(_ => new TimePickerHandler());
+			handlersCollection.AddHandler<Page>(_ => new PageHandler());
+			handlersCollection.AddHandler<WebView>(_ => new WebViewHandler());
+			handlersCollection.AddHandler<Border>(_ => new BorderHandler());
+			handlersCollection.AddHandler<IContentView>(_ => new ContentViewHandler());
+			handlersCollection.AddHandler<Shapes.Ellipse>(_ => new ShapeViewHandler());
+			handlersCollection.AddHandler<Shapes.Line>(_ => new LineHandler());
+			handlersCollection.AddHandler<Shapes.Path>(_ => new PathHandler());
+			handlersCollection.AddHandler<Shapes.Polygon>(_ => new PolygonHandler());
+			handlersCollection.AddHandler<Shapes.Polyline>(_ => new PolylineHandler());
+			handlersCollection.AddHandler<Shapes.Rectangle>(_ => new RectangleHandler());
+			handlersCollection.AddHandler<Shapes.RoundRectangle>(_ => new RoundRectangleHandler());
+			handlersCollection.AddHandler<Window>(_ => new WindowHandler());
+			handlersCollection.AddHandler<ImageButton>(_ => new ImageButtonHandler());
+			handlersCollection.AddHandler<IndicatorView>(_ => new IndicatorViewHandler());
+			handlersCollection.AddHandler<RadioButton>(_ => new RadioButtonHandler());
+			handlersCollection.AddHandler<RefreshView>(_ => new RefreshViewHandler());
+			handlersCollection.AddHandler<SwipeItem>(_ => new SwipeItemMenuItemHandler());
+			handlersCollection.AddHandler<SwipeView>(_ => new SwipeViewHandler());
 
 #pragma warning disable CA1416 //  'MenuBarHandler', MenuFlyoutSubItemHandler, MenuFlyoutSubItemHandler, MenuBarItemHandler is only supported on: 'ios' 13.0 and later
-			handlersCollection.AddHandler<MenuBar, MenuBarHandler>();
-			handlersCollection.AddHandler<MenuFlyoutSubItem, MenuFlyoutSubItemHandler>();
-			handlersCollection.AddHandler<MenuFlyoutSeparator, MenuFlyoutSeparatorHandler>();
-			handlersCollection.AddHandler<MenuFlyoutItem, MenuFlyoutItemHandler>();
-			handlersCollection.AddHandler<MenuBarItem, MenuBarItemHandler>();
+			handlersCollection.AddHandler<MenuBar>(_ => new MenuBarHandler());
+			handlersCollection.AddHandler<MenuFlyoutSubItem>(_ => new MenuFlyoutSubItemHandler());
+			handlersCollection.AddHandler<MenuFlyoutSeparator>(_ => new MenuFlyoutSeparatorHandler());
+			handlersCollection.AddHandler<MenuFlyoutItem>(_ => new MenuFlyoutItemHandler());
+			handlersCollection.AddHandler<MenuBarItem>(_ => new MenuBarItemHandler());
 #pragma warning restore CA1416
 
 #if WINDOWS || ANDROID || IOS || MACCATALYST || TIZEN
-			handlersCollection.AddHandler(typeof(ListView), typeof(Handlers.Compatibility.ListViewRenderer));
+			handlersCollection.AddHandler<ListView>(svc => new Handlers.Compatibility.ListViewRenderer(
+	#if ANDROID
+				svc.GetRequiredService<Android.Content.Context>()
+	#endif
+			));
 #if !TIZEN
-			handlersCollection.AddHandler(typeof(Cell), typeof(Handlers.Compatibility.CellRenderer));
-			handlersCollection.AddHandler(typeof(ImageCell), typeof(Handlers.Compatibility.ImageCellRenderer));
-			handlersCollection.AddHandler(typeof(EntryCell), typeof(Handlers.Compatibility.EntryCellRenderer));
-			handlersCollection.AddHandler(typeof(TextCell), typeof(Handlers.Compatibility.TextCellRenderer));
-			handlersCollection.AddHandler(typeof(ViewCell), typeof(Handlers.Compatibility.ViewCellRenderer));
-			handlersCollection.AddHandler(typeof(SwitchCell), typeof(Handlers.Compatibility.SwitchCellRenderer));
+	#if !WINDOWS
+			handlersCollection.AddHandler<Cell>(_ => new Handlers.Compatibility.CellRenderer());
+	#endif
+			handlersCollection.AddHandler<ImageCell>(_ => new Handlers.Compatibility.ImageCellRenderer());
+			handlersCollection.AddHandler<EntryCell>(_ => new Handlers.Compatibility.EntryCellRenderer());
+			handlersCollection.AddHandler<TextCell>(_ => new Handlers.Compatibility.TextCellRenderer());
+			handlersCollection.AddHandler<ViewCell>(_ => new Handlers.Compatibility.ViewCellRenderer());
+			handlersCollection.AddHandler<SwitchCell>(_ => new Handlers.Compatibility.SwitchCellRenderer());
 #endif
-			handlersCollection.AddHandler(typeof(TableView), typeof(Handlers.Compatibility.TableViewRenderer));
-			handlersCollection.AddHandler(typeof(Frame), typeof(Handlers.Compatibility.FrameRenderer));
+			handlersCollection.AddHandler<TableView>(svc => new Handlers.Compatibility.TableViewRenderer(
+	#if ANDROID
+				svc.GetRequiredService<Android.Content.Context>()
+	#endif
+			));
+			handlersCollection.AddHandler<Frame>(svc => new Handlers.Compatibility.FrameRenderer(
+	#if ANDROID
+				svc.GetRequiredService<Android.Content.Context>()
+	#endif
+			));
 #endif
 
 #if WINDOWS || MACCATALYST
-			handlersCollection.AddHandler(typeof(MenuFlyout), typeof(MenuFlyoutHandler));
+			handlersCollection.AddHandler<MenuFlyout>(_ => new MenuFlyoutHandler());
 #endif
 
 #if IOS || MACCATALYST
-			handlersCollection.AddHandler(typeof(NavigationPage), typeof(Handlers.Compatibility.NavigationRenderer));
-			handlersCollection.AddHandler(typeof(TabbedPage), typeof(Handlers.Compatibility.TabbedRenderer));
-			handlersCollection.AddHandler(typeof(FlyoutPage), typeof(Handlers.Compatibility.PhoneFlyoutPageRenderer));
+			handlersCollection.AddHandler<NavigationPage>(_ => new Handlers.Compatibility.NavigationRenderer());
+			handlersCollection.AddHandler<TabbedPage>(_ => new Handlers.Compatibility.TabbedRenderer());
+			handlersCollection.AddHandler<FlyoutPage>(_ => new Handlers.Compatibility.PhoneFlyoutPageRenderer());
 #endif
 
 #if ANDROID || IOS || MACCATALYST || TIZEN
-			handlersCollection.AddHandler<SwipeItemView, SwipeItemViewHandler>();
+			handlersCollection.AddHandler<SwipeItemView>(_ => new SwipeItemViewHandler());
 #if ANDROID || IOS || MACCATALYST
-			handlersCollection.AddHandler<Shell, ShellRenderer>();
+			handlersCollection.AddHandler<Shell>(_ => new ShellRenderer());
 #else
-			handlersCollection.AddHandler<Shell, ShellHandler>();
-			handlersCollection.AddHandler<ShellItem, ShellItemHandler>();
-			handlersCollection.AddHandler<ShellSection, ShellSectionHandler>();
+			handlersCollection.AddHandler<Shell>(_ => new ShellHandler());
+			handlersCollection.AddHandler<ShellItem>(_ => new ShellItemHandler());
+			handlersCollection.AddHandler<ShellSection>(_ => new ShellSectionHandler());
 #endif
 #endif
 #if WINDOWS || ANDROID || TIZEN
-			handlersCollection.AddHandler<NavigationPage, NavigationViewHandler>();
-			handlersCollection.AddHandler<Toolbar, ToolbarHandler>();
-			handlersCollection.AddHandler<FlyoutPage, FlyoutViewHandler>();
-			handlersCollection.AddHandler<TabbedPage, TabbedViewHandler>();
+			handlersCollection.AddHandler<NavigationPage>(_ => new NavigationViewHandler());
+			handlersCollection.AddHandler<Toolbar>(_ => new ToolbarHandler());
+			handlersCollection.AddHandler<FlyoutPage>(_ => new FlyoutViewHandler());
+			handlersCollection.AddHandler<TabbedPage>(_ => new TabbedViewHandler());
 #endif
 
 #if WINDOWS
-			handlersCollection.AddHandler<ShellItem, ShellItemHandler>();
-			handlersCollection.AddHandler<ShellSection, ShellSectionHandler>();
-			handlersCollection.AddHandler<ShellContent, ShellContentHandler>();
-			handlersCollection.AddHandler<Shell, ShellHandler>();
+			handlersCollection.AddHandler<ShellItem>(_ => new ShellItemHandler());
+			handlersCollection.AddHandler<ShellSection>(_ => new ShellSectionHandler());
+			handlersCollection.AddHandler<ShellContent>(_ => new ShellContentHandler());
+			handlersCollection.AddHandler<Shell>(_ => new ShellHandler());
 #endif
 			return handlersCollection;
 		}