… eng/BannedSymbols.txt
… src/Core/src/PublicAPI/net-ios/PublicAPI.Unshipped.txt
… src/Core/src/PublicAPI/net-maccatalyst/PublicAPI.Unshipped.txt
… src/Core/src/PublicAPI/net-tizen/PublicAPI.Unshipped.txt
… src/Core/src/PublicAPI/net-windows/PublicAPI.Unshipped.txt
… src/Core/src/PublicAPI/net/PublicAPI.Unshipped.txt
… src/Core/src/PublicAPI/netstandard/PublicAPI.Unshipped.txt
… src/Core/src/PublicAPI/netstandard2.0/PublicAPI.Unshipped.txt