@PureWeen 2023-08-23
@jonathanpeppers @mattleibow so one thing I realized about this PR is that this will disable our ability to do ctor injection with handlers. I was more hoping we were going to move to instantiating all the handlers via `DependencyInjection.ActivatorUtilities.CreateInstance` or service resolution vs moving to factories. 

https://github.com/dotnet/maui/issues/4129

I realize users could re-register the handler but that's a bit awkward. 

Will we eventually be able to move these back once the issue that regressed with runtime is fixed?