@jonathanpeppers 2023-08-23
> Will we eventually be able to move these back once the issue that regressed with runtime is fixed?

We should avoid DI features that use System.Reflection -- if we want a fast startup time. They might have a source generator one day, but that is some future .NET 9 thing.