@PureWeen 2023-08-23
> > this will disable our ability to do dependency injection with handlers
> 
> @PureWeen no, it shouldn't. There is an example how to do this in [e7812b0](https://github.com/dotnet/maui/commit/e7812b0f00ffebfe753cc8f35479aa7a4fbf6135):
> 
> https://github.com/dotnet/maui/blob/862fa07227bdad4a26175d785492f697eb37ab5e/src/Core/src/Hosting/HandlerMauiAppBuilderExtensions.cs#L19
> 
> If you want to do constructor injection, you just have to pass the parameters in.
> 
> Keep in mind customers won't have to do any of this in their apps. They can register things however they like.

yea, I was thinking from our performance analysis work in NET6 we'd mostly determined it didn't really matter. I was just hopeful we'd be able to avoid needing to be prescriptive about how to register. 

If this is something users need to be aware of in general, we might also look at updating our docs https://learn.microsoft.com/en-us/dotnet/architecture/maui/dependency-injection#registration

Or eventually adding to the queue of things that would be neat for a MAUI analyzer :-)